requirements.txt
# Core dependencies
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
tqdm>=4.66.0
omegaconf>=2.3.0      # Configuration management (YAML + CLI)

# API and async support
aiohttp>=3.9.0
requests>=2.31.0
anthropic>=0.18.0     # Claude API client

# Image quality assessment
image-quality>=1.2.7  # Preferred for Apple Silicon
# pybrisque>=0.2.5    # Alternative if image-quality fails

# Secure API key storage
keyring>=24.0.0

# Optional but recommended
python-dotenv>=1.0.0  # For .env file support

# Metadata embedding
piexif>=1.1.3        # JPEG EXIF manipulation
defusedxml>=0.7.1    # Safe XML parsing for XMP

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Optional format support (install if needed)
pillow-heif>=0.16.0  # HEIC/HEIF support (recommended for iPhone images)

# Event system
redis>=5.0.0         # Redis client for event persistence
asyncio-redis>=0.16.0  # Async Redis support